[⚡ code review] 타입 IBtn 분리 작업 취소
`IBtn`이 2회 이상 사용되었기 때문에 최초 types 폴더로 타입을 분리했지만, 추가 사용될 여지가 없기 때문에 분리하지 않는 것으로 코드 리뷰를 통해 의견 정리
