+++ src/components/common/Btn.tsx
@@ -1,5 +1,13 @@
 import styled from '@emotion/styled';
-import IBtn from '@/types/IBtn';
+
+interface IBtn {
+  type: 'button' | 'submit';
+  text: string;
+  isAble: boolean;
+  backgroundColor: string;
+  fontColor: string;
+  size: 'large' | 'small';
+}
 
 interface IBtnWrapper {
   btns: IBtn[];
