--- src/types/IBtn.ts
@@ -1,10 +1,0 @@
-interface IBtn {
-  type: 'button' | 'submit';
-  text: string;
-  isAble: boolean;
-  backgroundColor: string;
-  fontColor: string;
-  size: 'large' | 'small';
-}
-
-export default IBtn;
